server/requirements.txt
fastapi
uvicorn
pydantic
llama-cpp-python
language_tool_python
huggingface-hub
